{{ cookiecutter.project_name }}/composition_repository/djangocms_picture/requirements.txt
djangocms-picture==4.1.1
